refactor: inline table_sort functions into table_rows_view
The compareCells, sortRows, and cellPreview functions are only used
by the TableRowsView database component. Inlining them directly into
TableRowsView eliminates an unnecessary utility file and reduces
project complexity.
